refactor session launch config rendering
